+++ src/common/utils/form-iframes/change.ts
@@ -1,5 +1,5 @@
 import { eventsEmitter } from 'src/tokenization'
-import { EventListener } from '../form-events'
+import { EventListener, validation } from '../form-events'
 
 export function change() {
   const windowMessage = new EventListener(window.parent)
@@ -10,26 +10,15 @@ export function change() {
     const parentNode = document.querySelector(`#${data?.fieldType}`)
     if (!parentNode) return
 
-    if (type === 'validation') {
-      const isValid = data.valid
-      const isEmpty = data.empty
-      const isPotentiallyValid = data.potentialValid
-
-      if (isEmpty || isPotentiallyValid) {
-        parentNode?.classList.remove('malga-hosted-field-valid')
-        parentNode?.classList.remove('malga-hosted-field-invalid')
-        return
-      }
-
-      parentNode?.classList.toggle('malga-hosted-field-valid', isValid)
-      parentNode?.classList.toggle('malga-hosted-field-invalid', !isValid)
-
-      eventsEmitter.emit('validity', {
-        field: data.fieldType,
-        error: data.error,
-        valid: data.valid,
-        parent: parentNode,
+    if (type === 'cardTypeChanged') {
+      eventsEmitter.emit('cardTypeChanged', {
+        card: data.card,
+        parentNode: parentNode,
       })
+    }
+
+    if (type === 'validation') {
+      validation(parentNode, data)
     }
 
     if (type === 'focus' || type === 'blur') {
