+++ src/tokenization.ts
@@ -1,9 +1,8 @@
 import type { MalgaConfigurations } from 'src/common/interfaces'
 
 import { Tokenize } from './tokenize'
-import { loaded } from './common/utils'
+import { Events, loaded } from './common/utils'
 import { change } from './common/utils/form-iframes/change'
-import { Events } from './events'
 
 export const eventsEmitter = new Events()
